refactor: add custom hooks
